backend/requirements.txt
django==5.0.3
djangorestframework==3.15.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
PyJWT==2.8.0
python-dotenv==1.0.1
gunicorn==21.2.0
dj-database-url==2.1.0
requests==2.31.0

